Control.Error.Script

Description

Use this module if you like to write simple scripts with Text-based errors, but you prefer to use ExceptT to handle errors rather than Control.Exception.

import Control.Error

main = runScript $ do
    str <- scriptIO getLine
    n   <- tryRead "Read failed" str
    scriptIO $ print (n + 1)

The Script Monad

type Script = ExceptT Text IO

An IO action that can fail with a Text error message

runScript :: Script a -> IO a

Runs the Script monad

Prints the first error to stderr and exits with exitFailure

scriptIO :: MonadIO m => IO a -> ExceptT Text m a

scriptIO resembles lift, except it catches all exceptions and converts them to Text

Note that scriptIO is compatible with the Script monad.